--- a/modules/angular2/src/testing/async.dart
+++ b/modules/angular2/src/testing/async.dart
@@ -1,0 +1,4 @@
+// This symbol is not used on the Dart side. This exists just as a stub.
+async(Function fn) {
+  throw 'async() test wrapper not available for Dart.';
+}