--- a/modules/angular2/src/testing/testing_internal_core.dart
+++ b/modules/angular2/src/testing/testing_internal_core.dart
@@ -26,30 +26,17 @@ import 'package:angular2/src/core/reflection/reflection_capabilities.dart';
 import 'package:angular2/src/core/di/provider.dart' show bind;
 import 'package:angular2/src/facade/collection.dart' show StringMapWrapper;
 
+import 'async_test_completer.dart';
+export 'async_test_completer.dart' show AsyncTestCompleter;
+
 import 'test_injector.dart';
 export 'test_injector.dart' show inject;
 
 TestInjector _testInjector = getTestInjector();
-bool _isCurrentTestAsync;
-Future _currentTestFuture;
 bool _inIt = false;
 bool _initialized = false;
 List<dynamic> _platformProviders = [];
 List<dynamic> _applicationProviders = [];
-
-class AsyncTestCompleter {
-  final _completer = new Completer();
-
-  AsyncTestCompleter() {
-    _currentTestFuture = this.future;
-  }
-
-  void done() {
-    _completer.complete();
-  }
-
-  Future get future => _completer.future;
-}
 
 void setDartBaseTestProviders(List<dynamic> platform, List<dynamic> application) {
   _platformProviders = platform;
@@ -70,18 +57,15 @@ void testSetup() {
 
   gns.beforeEach(() {
     _testInjector.reset();
-    _currentTestFuture = null;
   });
 
   var completerProvider = bind(AsyncTestCompleter).toFactory(() {
     // Mark the test as async when an AsyncTestCompleter is injected in an it(),
     if (!_inIt) throw 'AsyncTestCompleter can only be injected in an "it()"';
-    _isCurrentTestAsync = true;
     return new AsyncTestCompleter();
   });
 
   gns.beforeEach(() {
-    _isCurrentTestAsync = false;
     _testInjector.addProviders([completerProvider]);
   });
 }
@@ -113,22 +97,16 @@ void beforeEachBindings(Function fn) {
 
 void beforeEach(fn) {
   testSetup();
-  if (fn is! FunctionWithParamTokens) fn =
-      new FunctionWithParamTokens([], fn, false);
-  gns.beforeEach(() {
-    _testInjector.execute(fn);
-  });
+  gns.beforeEach(fn);
 }
 
 void _it(gnsFn, name, fn) {
   testSetup();
-  if (fn is! FunctionWithParamTokens) fn =
-      new FunctionWithParamTokens([], fn, false);
   gnsFn(name, () {
     _inIt = true;
-    _testInjector.execute(fn);
+    var retVal = fn();
     _inIt = false;
-    if (_isCurrentTestAsync) return _currentTestFuture;
+    return retVal;
   });
 }