@juliemr 2016-04-26
A couple small comments but overall this looks great, much cleaner!

Can we add an example to https://github.com/angular/angular/blob/master/modules/angular2/test/testing/testing_public_spec.ts showing the use of `done` and `inject`? Something like


```js
it('should allow inject and done`, (done) => {
  inject([FancyService] (service) => {
    let count = 0;
    setInterval(() => {
      count++;
      if (count > 2) { done(); }
    }, 5);
  })();
});
```