@vikerman 2016-04-26
Test timing after all "it" methods are wrapped with (done)=> {}

JS(No change):
master: SUCCESS (27.341 secs / 23.577 secs)
PR: SUCCESS (26.694 secs / 23.059 secs)

CJS (Maybe little slower after change)
master: Finished in 12.52 seconds
PR: Finished in 13.01 seconds