@mary-poppins 2016-04-28
Merging PR #8257 on behalf of @hansl to branch [presubmit-hansl-pr-8257](https://github.com/angular/angular/tree/presubmit-hansl-pr-8257).